requirements.txt
python-dotenv
pyyaml
chardet
schedule